requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.5
django-analytical==3.0.0
django-braces==1.14.0
django-ckeditor==6.0.0
django-debug-toolbar==3.1.1
django-etc==1.2.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-model-utils==4.1.1
django-ranged-response==0.2.0
django-recaptcha3==0.4.0
django-simple-captcha==0.5.13
django-star-ratings==0.9.1
idna==2.10
mysqlclient @ file:///C:/Users/Govard/Desktop/testsite/mysqlclient-1.4.6-cp38-cp38-win_amd64.whl
Pillow==7.2.0
pysolr==3.9.0
pytils==0.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.10
